# server/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API documentation for the lifecycle management for Docker application. ",
    "version": "0.10.0",
    "title": "API Docs",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "scoady2@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://raw.githubusercontent.com/StephenCoady/lifecycle-management-for-docker/master/LICENCE"
    }
  },
  "basePath": "/api",
  "tags": [{
      "name": "container",
      "description": "Control containers on the host"
    },
    {
      "name": "image",
      "description": "Control images on the host"
    },
    {
      "name": "network",
      "description": "Control networks on the host"
    },
    {
      "name": "volume",
      "description": "Control volumes on the host"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/containers/running": {
      "get": {
        "tags": [
          "container"
        ],
        "summary": "get all running containers",
        "description": "",
        "operationId": "listRunningContainers",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Containers"
            }
          }
        }
      }
    },
    "/containers/all": {
      "get": {
        "tags": [
          "container"
        ],
        "summary": "get all containers",
        "description": "",
        "operationId": "listContainers",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Containers"
            }
          }
        }
      }
    },
    "/containers/{id}/": {
      "get": {
        "tags": [
          "container"
        ],
        "summary": "Find container by id",
        "description": "",
        "operationId": "listSpecificContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Container"
              }
            }
          },
          "404": {
            "description": "Container not found"
          }
        }
      }
    },
    "/containers/{id}/start": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Start any container on the host",
        "description": "",
        "operationId": "startContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Container started successfully"
          },
          "409": {
            "description": "Container cannot be started"
          }
        }
      }
    },
    "/containers/{id}/stop": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Stop any container on the host",
        "description": "",
        "operationId": "stopContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Container stopped successfully"
          },
          "409": {
            "description": "Container cannot be stopped"
          }
        }
      }
    },
    "/containers/{id}/pause": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Pause any container on the host",
        "description": "",
        "operationId": "pauseContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Container paused successfully"
          },
          "409": {
            "description": "Container cannot be paused"
          }
        }
      }
    },
    "/containers/{id}/unpause": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Unpause any container on the host",
        "description": "",
        "operationId": "unpauseContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Container unpaused successfully"
          },
          "409": {
            "description": "Container cannot be unpaused"
          }
        }
      }
    },
    "/containers/{id}/restart": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Restart any container on the host",
        "description": "",
        "operationId": "restartContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "container ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Container restarted successfully"
          },
          "409": {
            "description": "Container cannot be restarted"
          }
        }
      }
    },
    "/containers/{id}/create": {
      "post": {
        "tags": [
          "container"
        ],
        "summary": "Create a new container on the host",
        "description": "",
        "operationId": "createContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "container name",
            "required": true,
            "type": "string"
          },
          {
          "name": "body",
          "in": "body",
          "description": "The container to be created.",
          "required": true,
          "schema": {
            "$ref": "#/definitions/createContainer"
          }
        }],
        "responses": {
          "201": {
            "description": "Container created successfully"
          },
          "500": {
            "description": "Container cannot be created"
          }
        }
      }
    },
    "/containers/{id}/remove": {
      "delete": {
        "tags": [
          "container"
        ],
        "summary": "Removes a container",
        "description": "",
        "operationId": "removeContainer",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "Container id to delete",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "responses": {
          "200": {
            "description": "Container removed successfully"
          },
          "409": {
            "description": "Container cannot be removed"
          }
        }
      }
    },
    "/images/": {
      "get": {
        "tags": [
          "image"
        ],
        "summary": "List all images on the host",
        "description": "",
        "operationId": "listImages",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Images"
            }
          }
        }
      }
    },
    "/images/{id}/": {
      "get": {
        "tags": [
          "image"
        ],
        "summary": "List an image on the host",
        "description": "",
        "operationId": "listSpecificImage",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "image ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Image"
            }
          },
          "404": {
            "description": "No such image"
          }
        }
      },
      "delete": {
        "tags": [
          "image"
        ],
        "summary": "Delete an image on the host",
        "description": "",
        "operationId": "removeImage",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "image ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "Image removed successfully"
          },
          "409": {
            "description": "Image cannot be removed"
          }
        }
      }
    },
    "/images/{id}/history": {
      "get": {
        "tags": [
          "image"
        ],
        "summary": "List the history of an image on the host",
        "description": "",
        "operationId": "imageHistory",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "image ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "multi"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ImageHistory"
            }
          },
          "404": {
            "description": "No such image"
          }
        }
      }
    },
    "/images/pull": {
      "post": {
        "tags": [
          "image"
        ],
        "summary": "Pulls an image to the host",
        "description": "",
        "operationId": "pullImage",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "body",
          "in": "body",
          "description": "The image to be pulled to the host",
          "required": true,
          "schema": {
            "$ref": "#/definitions/PullImage"
          }
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "Cannot find the image requested"
          }
        }
      }
    },
    "/networks/": {
      "post": {
        "tags": [
          "network"
        ],
        "summary": "Create a new network on the host",
        "description": "",
        "operationId": "createNetwork",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
          "name": "body",
          "in": "body",
          "description": "The network to be created.",
          "required": true,
          "schema": {
            "required": [
              "Name"
            ],
            "properties": {
              "Name": {
                "type": "string",
                "example": "my-network"
              }
            }          }
        }],
        "responses": {
          "201": {
            "description": "Network created successfully"
          },
          "500": {
            "description": "Network cannot be created"
          }
        }
      }
    },
    "/networks/{id}": {
      "delete": {
        "tags": [
          "network"
        ],
        "summary": "Removes a network",
        "description": "",
        "operationId": "removeNetwork",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "Network id to delete",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "responses": {
          "200": {
            "description": "Network removed successfully"
          },
          "409": {
            "description": "Network cannot be removed"
          }
        }
      }
    },
    "/networks": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "List the networks on the host",
        "description": "",
        "operationId": "listNetworks",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Networks"
            }
          }
        }
      }
    },
    "/networks/{id}/": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "List a specific network on the host",
        "description": "",
        "operationId": "listSpecificNetwork",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "network ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "single"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Network"
            }
          },
          "404": {
            "description": "No such network"
          }
        }
      }
    },
    "/volumes/": {
      "post": {
        "tags": [
          "volume"
        ],
        "summary": "Create a new volume on the host",
        "description": "",
        "operationId": "createVolume",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
          "name": "body",
          "in": "body",
          "description": "The volume to be created.",
          "required": true,
          "schema": {
            "required": [
              "Name"
            ],
            "properties": {
              "Name": {
                "type": "string",
                "example": "my-volume"
              }
            }          }
        }],
        "responses": {
          "201": {
            "description": "Volume created successfully"
          },
          "500": {
            "description": "Volume cannot be created"
          }
        }
      }
    },
    "/volumes/{id}": {
      "delete": {
        "tags": [
          "volume"
        ],
        "summary": "Removes a volume",
        "description": "",
        "operationId": "removeVolume",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "Volume id to delete",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "responses": {
          "200": {
            "description": "Volume removed successfully"
          },
          "409": {
            "description": "Volume cannot be removed"
          }
        }
      }
    },
    "/volumes": {
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "List the volumes on the host",
        "description": "",
        "operationId": "listVolumes",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Volumes"
            }
          }
        }
      }
    },
    "/volumes/{id}/": {
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "List a specific volume on the host",
        "description": "",
        "operationId": "listSpecificVolume",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "description": "volume ID",
          "required": true,
          "type": "array",
          "items": {
            "type": "string"
          },
          "collectionFormat": "single"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Volume"
            }
          },
          "404": {
            "description": "No such volume"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Images": {
      "type": "object",
      "properties": {
        "images": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Containers": {
                "type": "string"
              },
              "Created": {
                "type": "integer"
              },
              "Id": {
                "type": "string"
              },
              "Labels": {
                "type": "string"
              },
              "ParentId": {
                "type": "integer"
              },
              "RepoDigests": {
                "type": "array",
                "items": {

                }
              },
              "RepoTags": {
                "type": "array",
                "items": {

                }
              },
              "SharedSize": {
                "type": "integer"
              },
              "Size": {
                "type": "integer"
              },
              "VirtualSize": {
                "type": "integer"
              }
            }
          }
        }
      }
    },
    "Containers": {
      "type": "object",
      "properties": {
        "containers": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Id": {
                "type": "string"
              },
              "Names": {
                "type": "array",
                "items": {}
              },
              "Image": {
                "type": "string"
              },
              "ImageID": {
                "type": "string"
              },
              "Command": {
                "type": "string"
              },
              "Created": {
                "type": "integer"
              },
              "Ports": [],
              "Labels": {},
              "State": {
                "type": "string"
              },
              "Status": {
                "type": "string"
              },
              "HostConfig": {
                "type": "object"
              },
              "NetworkSettings": {
                "type": "object"
              },
              "Mounts": {
                "type": "array",
                "items": {}
              }
            }
          }
        }
      }
    },
    "Image": {
      "type": "object",
      "properties": {
        "image": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "string"
            },
            "RepoTags": {
              "type": "string"
            },
            "RepoDigests": {
              "type": "string"
            },
            "Parent": {
              "type": "string"
            },
            "Comment": {
              "type": "string"
            },
            "Created": {
              "type": "string"
            },
            "Container": {
              "type": "string"
            },
            "ContainerConfig": {
              "type": "object"
            },
            "DockerVersion": {
              "type": "string"
            },
            "Author": {
              "type": "string"
            },
            "Config": {
              "type": "object"
            },
            "Architecture": {
              "type": "string"
            },
            "Os": {
              "type": "string"
            },
            "Size": {
              "type": "integer"
            },
            "VirtualSize": {
              "type": "integer"
            },
            "GraphDriver": {
              "type": "object"
            },
            "RootFS": {
              "type": "object"
            }
          }
        }
      }
    },
    "ImageHistory": {
      "type": "object",
      "properties": {
        "image": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Id": {
                "type": "object"
              },
              "Created": {
                "type": "integer"
              },
              "CreatedBy": {
                "type": "string"
              },
              "Tags": {
                "type": "array",
                "properties": {}
              },
              "Size": {
                "type": "integer"
              },
              "Comment": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "Container": {
      "type": "object",
      "properties": {
        "container": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "string"
            },
            "Created": {
              "type": "string"
            },
            "Path": {
              "type": "string"
            },
            "Args": {
              "type": "string"
            },
            "State": {
              "type": "object"
            },
            "Image": {
              "type": "string"
            },
            "ResolvConfPath": {
              "type": "string"
            },
            "HostnamePath": {
              "type": "string"
            },
            "HostsPath": {
              "type": "string"
            },
            "LogPath": {
              "type": "string"
            },
            "Name": {
              "type": "string"
            },
            "RestartCount": {
              "type": "integer"
            },
            "Driver": {
              "type": "string"
            },
            "MountLabel": {
              "type": "string"
            },
            "ProcessLabel": {
              "type": "string"
            },
            "AppArmorProfile": {
              "type": "string"
            },
            "ExecIDs": {
              "type": "string"
            },
            "HostConfig": {
              "type": "object"
            },
            "GraphDriver": {
              "type": "object"
            },
            "Mounts": {
              "type": "object"
            },
            "Config": {
              "type": "object"
            },
            "NetworkSettings": {
              "type": "object"
            }
          }
        }
      }
    },
    "Networks": {
      "type": "object",
      "properties": {
        "networks": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Name": {
                "type": "string"
              },
              "Id": {
                "type": "string"
              },
              "Created": {
                "type": "string"
              },
              "Scope": {
                "type": "string"
              },
              "Driver": {
                "type": "string"
              },
              "EnableIPv6": {
                "type": "boolean"
              },
              "IPAM": {
                "type": "object"
              },
              "Internal": {
                "type": "boolean"
              },
              "Attachable": {
                "type": "boolean"
              },
              "Containers": {
                "type": "object"
              },
              "Options": {
                "type": "object"
              },
              "Labels": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "Network": {
      "type": "object",
      "properties": {
        "network": {
          "type": "object",
          "properties": {
            "Name": {
              "type": "string"
            },
            "Id": {
              "type": "string"
            },
            "Created": {
              "type": "string"
            },
            "Scope": {
              "type": "string"
            },
            "Driver": {
              "type": "string"
            },
            "EnableIPv6": {
              "type": "boolean"
            },
            "IPAM": {
              "type": "object"
            },
            "Internal": {
              "type": "boolean"
            },
            "Attachable": {
              "type": "boolean"
            },
            "Containers": {
              "type": "object"
            },
            "Options": {
              "type": "object"
            },
            "Labels": {
              "type": "object"
            }
          }
        }
      }
    },
    "Volumes": {
      "type": "object",
      "properties": {
        "volumes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "Driver": {
                "type": "string"
              },
              "Labels": {
                "type": "string"
              },
              "Mountpoint": {
                "type": "string"
              },
              "Name": {
                "type": "string"
              },
              "Options": {
                "type": "string"
              },
              "Scope": {
                "type": "boolean"
              }
            }
          }
        }
      }
    },
    "Volume": {
      "type": "object",
      "properties": {
        "volume": {
          "type": "object",
          "properties": {
            "Driver": {
              "type": "string"
            },
            "Labels": {
              "type": "string"
            },
            "Mountpoint": {
              "type": "string"
            },
            "Name": {
              "type": "string"
            },
            "Options": {
              "type": "string"
            },
            "Scope": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "PullImage": {
      "required": [
        "repo",
        "name"
      ],
      "properties": {
        "repo": {
          "type": "string",
          "example": "library"
        },
        "name": {
          "type": "string",
          "example": "alpine"
        },
        "tag": {
          "type": "string",
          "example": "3.1"
        }
      }
    },
    "createContainer": {
      "required": [
        "Image"
      ],
      "properties": {
        "Image": {
          "type": "string",
          "example": "alpine"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}